requirements.txt
pyserial
pynmea2
datetime
